fix(pagination): a page number is bounded, and the bound is published (#1163)
Closes F34 — with one correction to what the audit claimed.

**The two real overflows are in Studio.** The audit trail and the data browser
parse `page` themselves rather than going through the framework's binder, and
slice in int arithmetic. `(page - 1) * size` wraps negative from page 42,949,674
upward, so the audit trail's window became `subList(-100, -50)` — which throws
for any log size, including an empty one, and rendered as TQL-ROUTE-5000. The
data browser's offset wrapped into `setMaxRows` with a negative row count. Both
widen before the multiply. StudioServiceTest errors on the old code, which is the
proof.

**The binder's own bound is a contract change, not a behaviour change, and the
commit says so rather than claiming the audit's 500.** F34 says a crafted `?page=`
answers 500 on any public list route. Measured on a build with the bound removed,
a declarative offset route already answers 400 with TQL-FIELD-2001 for 2^31,
3,000,000,000 and Long.MAX_VALUE alike, and 200 at Integer.MAX_VALUE — behaviour
identical to the bounded build. I could not produce an input whose HTTP answer
differs, so the end-to-end cases are labelled as pins on behaviour that already
holds, not as regression tests.

What the bound does add is worth having on its own terms: the ceiling is stated
in one place instead of being an emergent property of three products, and
OpenApiGenerator publishes it as the `page` parameter's `maximum` beside the one
`size` already published. A framework-owned parameter whose refusal exists only
in code is a contract the generated document lies about, and openapi.json is a
shipped artifact — an operator's kept baseline will see `maximum` appear on
`page`, and `openapiSha256` moves in release evidence for every app that
regenerates.

The ceiling is flat rather than derived from the declared size: with the page at
most Integer.MAX_VALUE and size an int, the largest product any of these
expressions can form stays inside long, so one constant removes the overflow from
every site instead of a divisor re-derived per route. The `size` call site keeps
its old behaviour explicitly — it is already clamped by Math.min against a
declared maxSize the contract publishes, and a page-shaped bound there would
silently refuse a large `?size=` that is harmless today.

+++ tesseraql-studio/src/main/java/io/tesseraql/studio/AuditTrail.java
@@ -82,8 +82,12 @@ final class AuditTrail {
         Comparator<AuditEntry> cmp = comparator(state.key());
         all.sort(state.descending() ? cmp.reversed() : cmp);
         int total = all.size();
-        int from = Math.min((p - 1) * size, total);
-        int to = Math.min(from + size, total);
+        // Widened to long before the multiply: this route parses its own page number rather
+        // than going through PageBinder, and in int arithmetic (p - 1) * size wraps negative
+        // from page 42,949,674 upward. At ?page=2147483647 the window became subList(-100, -50),
+        // which throws for any log size including an empty one, rendered as an internal error.
+        int from = (int) Math.min((long) (p - 1) * size, total);
+        int to = (int) Math.min((long) from + size, total);
         return new AuditPage(List.copyOf(all.subList(from, to)), p, size, total);
     }
